+++ .vscode/settings.json
@@ -3,5 +3,8 @@
   "prettier.semi": false,
   "prettier.singleQuote": true,
   "prettier.tabWidth": 2,
-  "prettier.trailingComma": "es5"
+  "prettier.trailingComma": "es5",
+  "files.exclude": {
+    "node_modules": true
+  }
 }
